Moved all extern C functions into new separate crate
