+++ src/katsuba/src/cmd/nav.rs
@@ -1,15 +1,8 @@
-use std::{
-    ffi::{CStr},
-    path::PathBuf,
-};
-
-use libc::{c_char};
-
 use clap::{Args, Subcommand, ValueEnum};
 use katsuba_nav::{NavigationGraph, ZoneNavigationGraph};
 
 use super::Command;
-use crate::cli::{helpers, Bias, InputsOutputs, Processor, HYPHEN};
+use crate::cli::{helpers, Bias, InputsOutputs, Processor};
 
 /// Subcommand for working with NAV data.
 #[derive(Debug, Args)]
@@ -50,7 +43,7 @@ impl Command for Nav {
     }
 }
 
-fn deserialize_nav(args: InputsOutputs) -> eyre::Result<()> {
+pub fn deserialize_nav(args: InputsOutputs) -> eyre::Result<()> {
     let (inputs, outputs) = args.evaluate("de.json")?;
     Processor::new(Bias::Current)?
         .read_with(|r, _| NavigationGraph::parse(r).map_err(Into::into))
@@ -58,76 +51,10 @@ fn deserialize_nav(args: InputsOutputs) -> eyre::Result<()> {
         .process(inputs, outputs)
 }
 
-fn deserialize_zonenav(args: InputsOutputs) -> eyre::Result<()> {
+pub fn deserialize_zonenav(args: InputsOutputs) -> eyre::Result<()> {
     let (inputs, outputs) = args.evaluate("de.json")?;
     Processor::new(Bias::Current)?
         .read_with(|r, _| ZoneNavigationGraph::parse(r).map_err(Into::into))
         .write_with(helpers::write_as_json)
         .process(inputs, outputs)
-}
-
-#[no_mangle]
-pub extern "C" fn nav_deserialize(
-    input: *const c_char,
-    output: *const c_char,
-) -> bool {
-    let default_path = PathBuf::from(HYPHEN);
-
-    let rust_input = if input.is_null() {
-        return false
-    } else {
-        match unsafe { CStr::from_ptr(input) }.to_str() {
-            Ok(rust_str) => rust_str.to_owned(),
-            Err(_) => return false,
-        }
-    };
-
-    let rust_output = if output.is_null() {
-        default_path
-    } else {
-        match unsafe { CStr::from_ptr(output) }.to_str() {
-            Ok(rust_str) => PathBuf::from(rust_str),
-            Err(_) => default_path,
-        }
-    };
-
-    let io = InputsOutputs {
-        input: rust_input,
-        output: rust_output,
-    };
-
-    deserialize_nav(io).is_ok()
-}
-
-#[no_mangle]
-pub extern "C" fn zonenav_deserialize(
-    input: *const c_char,
-    output: *const c_char,
-) -> bool {
-    let default_path = PathBuf::from(HYPHEN);
-
-    let rust_input = if input.is_null() {
-        return false
-    } else {
-        match unsafe { CStr::from_ptr(input) }.to_str() {
-            Ok(rust_str) => rust_str.to_owned(),
-            Err(_) => return false,
-        }
-    };
-
-    let rust_output = if output.is_null() {
-        default_path
-    } else {
-        match unsafe { CStr::from_ptr(output) }.to_str() {
-            Ok(rust_str) => PathBuf::from(rust_str),
-            Err(_) => default_path,
-        }
-    };
-
-    let io = InputsOutputs {
-        input: rust_input,
-        output: rust_output,
-    };
-
-    deserialize_zonenav(io).is_ok()
 }
