+++ src/katsuba/src/cmd/op.rs
@@ -1,20 +1,14 @@
-use std::{
-    ffi::CStr,
-    path::PathBuf,
-    sync::Arc
-};
-
-use libc::{c_char};
+use std::{path::PathBuf, sync::Arc};
 
 use clap::{Args, Subcommand};
-use katsuba_object_property::serde::{self, SerializerOptions};
+use katsuba_object_property::serde;
 use katsuba_types::{PropertyFlags, TypeList};
 
 use super::Command;
-use crate::cli::{helpers, Bias, InputsOutputs, Processor, HYPHEN};
+use crate::cli::{helpers, Bias, InputsOutputs, Processor};
 
-mod guess;
-mod utils;
+pub mod guess;
+pub mod utils;
 
 pub const DEFAULT_FLAGS: u32 = 0;
 pub const DEFAULT_MASK: u32 = 0x18;
@@ -141,10 +135,10 @@ impl Command for ObjectProperty {
     }
 }
 
-fn deserialize(
+pub fn deserialize(
     args: InputsOutputs,
     type_list: Arc<TypeList>,
-    mut options: SerializerOptions,
+    mut options: serde::SerializerOptions,
     ignore_unknown_types: bool,
 ) -> eyre::Result<()> {
     let (inputs, outputs) = args.evaluate("de.xml")?;
@@ -171,120 +165,4 @@ fn deserialize(
         })
         .write_with(helpers::write_as_json)
         .process(inputs, outputs)
-}
-
-fn get_type_lists_from_c(type_lists: *const *const c_char) -> eyre::Result<Arc<TypeList>> {
-    let mut type_list_paths = Vec::new();
-    let mut i = 0;
-    loop {
-        let current_type_list = unsafe { *type_lists.add(i) };
-        if current_type_list.is_null() {
-            break // Null pointer indicates end of array
-        }
-
-        match unsafe { CStr::from_ptr(current_type_list) }.to_str() {
-            Ok(rust_str) => type_list_paths.push(PathBuf::from(rust_str)),
-            Err(_) => continue,
-        };
-
-        i += 1;
-    }
-
-    match utils::merge_type_lists(type_list_paths) {
-        Ok(combined_type_list) => Ok(Arc::new(combined_type_list)),
-        Err(report) => Err(report),
-    }
-}
-
-#[no_mangle]
-pub extern "C" fn op_deserialize(
-    input: *const c_char,
-    output: *const c_char,
-    type_lists: *const *const c_char,
-    flags: u32,
-    mask: u32,
-    shallow: bool,
-    manual_compression: bool,
-    djb2_only: bool,
-    ignore_unknown_types: bool,
-) -> bool {
-    let default_path = PathBuf::from(HYPHEN);
-
-    if input.is_null() || type_lists.is_null() {
-        return false
-    }
-
-    // Create the InputsOutputs
-    let rust_input = match unsafe { CStr::from_ptr(input) }.to_str() {
-        Ok(rust_str) => rust_str.to_owned(),
-        Err(_) => return false,
-    };
-
-    let rust_output = if output.is_null() {
-        default_path
-    } else {
-        match unsafe { CStr::from_ptr(output) }.to_str() {
-            Ok(rust_str) => PathBuf::from(rust_str),
-            Err(_) => default_path,
-        }
-    };
-
-    let io = InputsOutputs {
-        input: rust_input,
-        output: rust_output,
-    };
-
-    // Create the type_list
-    let type_list = match get_type_lists_from_c(type_lists) {
-        Ok(list) => list,
-        Err(_) => return false,
-    };
-
-    // Set the options
-    let options = serde::SerializerOptions {
-        flags: serde::SerializerFlags::from_bits_truncate(flags),
-        property_mask: PropertyFlags::from_bits_truncate(mask),
-        shallow: shallow,
-        manual_compression: manual_compression,
-        djb2_only: djb2_only,
-        ..Default::default()
-    };
-
-    deserialize(io, type_list, options, ignore_unknown_types).is_ok()
-}
-
-#[no_mangle]
-pub extern "C" fn op_guess(
-    path: *const c_char,
-    type_lists: *const *const c_char,
-    flags: u32,
-    mask: u32,
-    shallow: bool,
-    manual_compression: bool,
-    djb2_only: bool,
-    quiet: bool,
-) -> bool {
-
-    // Set the options
-    let options = serde::SerializerOptions {
-        flags: serde::SerializerFlags::from_bits_truncate(flags),
-        property_mask: PropertyFlags::from_bits_truncate(mask),
-        shallow: shallow,
-        manual_compression: manual_compression,
-        djb2_only: djb2_only,
-        ..Default::default()
-    };
-
-    // Create the type_list
-    let type_list = match get_type_lists_from_c(type_lists) {
-        Ok(list) => list,
-        Err(_) => return false,
-    };
-
-    let rust_path = match unsafe { CStr::from_ptr(path) }.to_str() {
-        Ok(rust_str) => PathBuf::from(rust_str),
-        Err(_) => return false,
-    };
-
-    guess::guess(options, type_list, rust_path, quiet).is_ok()
 }
