+++ src/katsuba/src/cmd/wad.rs
@@ -1,17 +1,14 @@
 use std::{
-    ffi::{CStr},
     fs,
     path::{Path, PathBuf},
 };
-
-use libc::{c_char};
 
 use clap::{Args, Subcommand};
 use eyre::Context;
 use katsuba_wad::{Archive, ArchiveBuilder};
 
 use super::Command;
-use crate::cli::{Bias, InputsOutputs, Processor, Reader, HYPHEN};
+use crate::cli::{Bias, InputsOutputs, Processor, Reader};
 
 mod extract;
 
@@ -79,7 +76,7 @@ impl Command for Wad {
     }
 }
 
-fn wad_pack(
+pub fn wad_pack(
     input: PathBuf,
     flags: u8,
     output: Option<PathBuf>,
@@ -128,7 +125,7 @@ fn wad_pack(
     Ok(())
 }
 
-fn wad_unpack(
+pub fn wad_unpack(
     args: InputsOutputs
 ) -> Result<(), eyre::Error> {
     let (inputs, outputs) = args.evaluate("")?;
@@ -143,64 +140,4 @@ fn wad_unpack(
         })
         .write_with(extract::extract_archive)
         .process(inputs, outputs)
-}
-
-#[no_mangle]
-pub extern "C" fn wad_pack_c(
-    input: *const c_char,
-    flags: u8,
-    output: *const c_char,
-) -> bool {
-    let input_path = if input.is_null() {
-        return false
-    } else {
-        match unsafe { CStr::from_ptr(input) }.to_str() {
-            Ok(rust_str) => PathBuf::from(rust_str),
-            Err(_) => return false,
-        }
-    };
-
-    let output_path = if output.is_null() {
-        None
-    } else {
-        match unsafe { CStr::from_ptr(output) }.to_str() {
-            Ok(rust_str) => Some(PathBuf::from(rust_str)),
-            Err(_) => None,
-        }
-    };
-
-    wad_pack(input_path, flags, output_path).is_ok()
-}
-
-#[no_mangle]
-pub extern "C" fn wad_unpack_c(
-    input: *const c_char,
-    output: *const c_char,
-) -> bool {
-    let rust_input = if input.is_null() {
-        return false
-    } else {
-        match unsafe { CStr::from_ptr(input) }.to_str() {
-            Ok(rust_str) => rust_str.to_owned(),
-            Err(_) => return false,
-        }
-    };
-
-    let default_path = PathBuf::from(HYPHEN);
-
-    let rust_output = if output.is_null() {
-        default_path
-    } else {
-        match unsafe { CStr::from_ptr(output) }.to_str() {
-            Ok(rust_str) => PathBuf::from(rust_str),
-            Err(_) => default_path,
-        }
-    };
-
-    let io = InputsOutputs {
-        input: rust_input,
-        output: rust_output,
-    };
-
-    wad_unpack(io).is_ok()
 }
